Break out Nav components; Move <Menu> to palette
